Modified the get neighbor methods in the graph and fixed the vertex bug. Now when a vertex is created without colors, the vertex's bit_vector will be all 1's so that it can accurately be anded to see which colors are contained in the graph. Required that the number of colors needed to be known by the Graph and vertex.

+++ inc/color_manager.h
@@ -65,7 +65,7 @@ class ColorManager {
 		vector<string> getColorNames(bit_vector colors);
 
         /// Returns the number of colors in the ColorManager
-        unsigned int getNumColors();
+        size_t getNumColors();
 
     private:
         /// The pointer to the file handle of the color file
@@ -80,7 +80,7 @@ class ColorManager {
 
         /// numColors is the total number of colors added.
         /// It also determines the ID of each Color.
-        unsigned int numColors;
+        size_t numColors;
 
 };
